requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0

# AI providers
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# Database for stats
aiosqlite>=0.20.0

# Telegram bot
aiogram>=3.4.0
httpx>=0.27.0
